requirements.txt

fastapi
uvicorn
pandas
numpy
openpyxl
qdrant-client
scikit-learn
openai
python-dotenv
joblib
